src/ComplexNumberTest.java
import org.junit.jupiter.api.Assertions;
import org.junit.jupiter.api.Test;

import java.util.ArrayList;

class ComplexNumberTest {

    @Test
    void whenInstantiatedInPolarForm_thenHasCorrectRectangularCoords() {
        ComplexNumber z1 = ComplexNumber.polar(1, 0);
        Assertions.assertEquals(1, z1.realPart(), 0.01);
        Assertions.assertEquals(0, z1.imaginaryPart(), 0.01);

        ComplexNumber z2 = ComplexNumber.polar(1, Math.PI / 4);
        Assertions.assertEquals(Math.round(Math.sqrt(2) / 2), Math.round(z2.realPart()));
        Assertions.assertEquals(Math.round(Math.sqrt(2) / 2), Math.round(z2.imaginaryPart()));

        ComplexNumber z3 = ComplexNumber.polar(1, 3 * Math.PI / 4);
        Assertions.assertEquals(Math.round(-Math.sqrt(2) / 2), Math.round(z3.realPart()));
        Assertions.assertEquals(Math.round(Math.sqrt(2) / 2), Math.round(z3.imaginaryPart()));

        ComplexNumber z4 = ComplexNumber.polar(1, 5 * Math.PI / 4);
        Assertions.assertEquals(Math.round(-Math.sqrt(2) / 2), Math.round(z4.realPart()));
        Assertions.assertEquals(Math.round(-Math.sqrt(2) / 2), Math.round(z4.imaginaryPart()));

        ComplexNumber z5 = ComplexNumber.polar(13, 5 * Math.PI / 4);
        Assertions.assertEquals(Math.round(-13 * Math.sqrt(2) / 2), Math.round(z5.realPart()));
        Assertions.assertEquals(Math.round(-13 * Math.sqrt(2) / 2), Math.round(z5.imaginaryPart()));

        ComplexNumber z6 = ComplexNumber.polar(1, 3 * (Math.PI / 2));
        Assertions.assertEquals(0, Math.round(z6.realPart()));
        Assertions.assertEquals(-1, z6.imaginaryPart(), 0.01);
    }
    @Test
    void product() {
    }

    @Test
    void testModulus() {
        Assertions.assertEquals(5, ComplexNumber.rectangular(3,4).modulus(), 0.001);
        Assertions.assertEquals(10, ComplexNumber.polar(10, 0).modulus(), 0.001);
        Assertions.assertEquals(39, ComplexNumber.rectangular(15,36).modulus(), 0.001);
        Assertions.assertEquals(0, ComplexNumber.polar(0, 0).modulus(), 0.001);
//        Assertions.assertEquals(18*Math.sqrt(626)/5, ComplexNumber.rectangular(3.6, 90).modulus(), 0.001);
    }

    @Test
    void testSqrt() {
        assert(ComplexNumber.rectangular(1,2).equals(ComplexNumber.rectangular(-3,4).sqrt()));
        assert(ComplexNumber.rectangular(0,3).equals(ComplexNumber.rectangular(-9,0).sqrt()));
        assert(ComplexNumber.rectangular(Math.sqrt(2),0)).equals((ComplexNumber.rectangular(2,0).sqrt()));
        assert(ComplexNumber.rectangular(8,-6).equals(ComplexNumber.rectangular(28,-96).sqrt()));
        assert(ComplexNumber.rectangular(24,63).equals(ComplexNumber.rectangular(-3393,3024).sqrt()));
        assert(ComplexNumber.rectangular(0.24,0.5).equals(ComplexNumber.rectangular(-0.1924,0.24).sqrt()));
        //Note tried a test with -2-7i but this will never happen when returning +ve sqrt
        assert(ComplexNumber.rectangular(2,7).equals(ComplexNumber.rectangular(-45,28).sqrt()));
    }

    @Test
    void testEquals(){
        ComplexNumber a = ComplexNumber.rectangular(3,4);
        assert(!ComplexNumber.rectangular(1,2).equals(a));
        ComplexNumber b = ComplexNumber.rectangular(4.6,1);
        assert(b.equals(ComplexNumber.rectangular(4.6,1)));
        assert(!b.equals(new ArrayList<String>()));
//        Problem with rounding when comparing - these 2 are identical but doesn't match due to way of implementation
//        System.out.println(a.toString() + ComplexNumber.polar(5, Math.atan(4.0/3.0)).toString());
//        assert(a.equals(ComplexNumber.polar(5, Math.atan(4.0/3.0))));
                //need to look into
    }

    @Test
    void whenInstantiatedInRectangularForm_thenHasCorrectArgument() {
        ComplexNumber z1 = ComplexNumber.rectangular(1, 1);
        Assertions.assertEquals(Math.PI / 4, z1.argument(), 0.001);

        ComplexNumber z2 = ComplexNumber.rectangular(-1, 1);
        Assertions.assertEquals(3 * Math.PI / 4, z2.argument(), 0.001);

        ComplexNumber z3 = ComplexNumber.polar(1, 3 * Math.PI / 2);
        Assertions.assertEquals(3 * Math.PI / 2, z3.argument(), 0.001);
    }


}